Created Search Component which is now working with the Resutls and Map components
