+++ config/config.json
@@ -1,7 +1,7 @@
 {
     "development": {
         "username": "root",
-        "password": "password",
+        "password": "hadary83",
         "database": "tukio_db",
         "host": "127.0.0.1",
         "dialect": "mysql"
